addColumn window WIP

+++ src/public/home/src/services/columnService.ts
@@ -1,5 +1,6 @@
 declare const tippy: typeof import('tippy.js').default;
 
+import { confirmPopup, notification } from '../../../utils/notification.js';
 import { Column } from '../types/column.js';
 import { UI } from '../UI/UI.js';
 
@@ -7,8 +8,10 @@ export const columnService = {
     columnMenuOptions: [
         {
             name: 'Delete column',
-            event: function (column: Column) {
-                // DEBUG
+            event: async function (column: Column) {
+                const confirmation = await confirmPopup();
+                if (!confirmation) return;
+
                 console.log("Deleting column " + column.name);
                 columnService.deleteColumn(column);
             }
@@ -73,8 +76,40 @@ export const columnService = {
         });
     },
 
+    async addColumn(name: string): Promise<Column | void> {
+        try {
+            const res = await fetch("/api/project/add-column", {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json",
+                },
+                body: JSON.stringify(
+                    { project_id: UI.activeProject!.id, name: name }
+                ),
+            });
+
+            const rJ = await res.json();
+
+            const column = new Column(
+                rJ.id,
+                rJ.project_id,
+                rJ.name,
+                []
+            );
+
+            // Add new column to activeProject columns and redraw
+            UI.activeProject!.columns.push(column);
+            UI.draw(UI.activeProject);
+
+            return column;
+        } catch (e) {
+            notification("Error occurred while adding the column - try again later!", "error");
+            return;
+        }
+    },
+
     deleteColumn(column: Column) {
-        UI.activeProject.columns = UI.activeProject.columns.filter(col => col.id !== column.id);
-        UI.draw(UI.activeProject);
+        UI.activeProject!.columns = UI.activeProject!.columns.filter(col => col.id !== column.id);
+        UI.draw(UI.activeProject!);
     },
 }
